Small reorganization of project
