+++ src/components/Content/Editor/scripts/DefaultTheme.ts
